Find most unique word and write some tests
Decoder currently finds the most unique word in the code. Next step is to find a list of words which match the coded layout of the most unique word.
Also started writing unit tests for private functions. Will be adding tests as more functions are added. Currently need to add tests for the |most_variety| function